Move ManageSpaceActivity.kt into managespace.ui
