+++ mobile/src/main/AndroidManifest.xml
@@ -16,7 +16,7 @@
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
-        android:manageSpaceActivity="com.boswelja.devicemanager.managespace.ManageSpaceActivity"
+        android:manageSpaceActivity="com.boswelja.devicemanager.managespace.ui.ManageSpaceActivity"
         android:resizeableActivity="true"
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
@@ -24,7 +24,7 @@
         tools:ignore="UnusedAttribute">
 
         <activity
-            android:name=".managespace.ManageSpaceActivity"
+            android:name=".managespace.ui.ManageSpaceActivity"
             android:label="@string/manage_space_title" />
 
         <activity android:name="com.boswelja.devicemanager.main.MainActivity">
