+++ mobile/src/main/java/com/boswelja/devicemanager/managespace/ui/ManageSpaceActivity.kt
@@ -5,7 +5,7 @@
  * This file, and any part of the Wearable Extensions app/s cannot be copied and/or distributed
  * without permission from Jack Boswell (boswelja) <boswela@outlook.com>
  */
-package com.boswelja.devicemanager.managespace
+package com.boswelja.devicemanager.managespace.ui
 
 import android.app.ActivityManager
 import android.content.Context
